IDOS->Seek() --> IDOS->ChangeFilePosition().

+++ dopus4-os4/trunk/program-os4/dopusproto.h
@@ -331,7 +331,7 @@ struct dopusfiletype *checkfiletype(char *, int, int);
 int checkfiletypefunc(char *, int);
 int dochecktype(struct dopusfiletype *, char *, int, struct FileInfoBlock *);
 int checktypechars(int, char *, int);
-int typesearch(int, char *, int, char *, int);
+int64 typesearch(int, char *, int, char *, int);
 int dorequest(struct DOpusSimpleRequest *, char *, char **, int *, struct Window *);
 int searchbuffer(char *, int, char *, int, int);
